Back-end completed

+++ scripts/truck.js
@@ -1,4 +1,4 @@
-(function (window) {
+(function(window) {
   'use strict';
   var App = window.App || {};
 
@@ -7,17 +7,21 @@
     this.db = db;
   }
 
-  Truck.prototype.createOrder = function (order) {
+  Truck.prototype.createOrder = function(order) {
     console.log('Adding order for ' + order.emailAddress);
     this.db.add(order.emailAddress, order);
   };
 
-  Truck.prototype.deliverOrder = function (customerId) {
-    console.log('Delivering order for ' + customerId);
-    this.db.remove(customerId);
+  Truck.prototype.deliverOrder = function(customerId) {
+    if (customerId !== undefined) {
+      console.log('Delivering order for ' + customerId);
+      this.db.remove(customerId);
+    } else {
+      console.log('Missing customer ID');
+    }
   };
 
-  Truck.prototype.printOrders = function () {
+  Truck.prototype.printOrders = function() {
     var customerIdArray = Object.keys(this.db.getAll());
 
     console.log('Truck #' + this.truckId + ' has pending orders:');
@@ -26,7 +30,7 @@
     }.bind(this))
   };
 
-    App.Truck = Truck;
-    window.App = App;
+  App.Truck = Truck;
+  window.App = App;
 
-}) (window);
+})(window);
